Fix ClassCastException in longListToArray and visitArray not-yet-supported
Agent-Logs-Url: https://github.com/pfirmstone/Pack200.net/sessions/7e494435-4114-4562-aeac-1737ab28a43e

+++ pack/src/main/java/org/apache/harmony/pack200/ClassBands.java
@@ -1071,7 +1071,7 @@ class ClassBands extends BandSet {
             if(latestCodeFlag == (1 << 2) && latestLocalVariableTableN == 0) {
                 codeLocalVariableTableN.remove(codeLocalVariableTableN.size() - 1);
                 codeFlags.remove(codeFlags.size() - 1);
-                codeFlags.add(Integer.valueOf(0));
+                codeFlags.add(Long.valueOf(0L));
             }
         }
     }
